Migrate token refresh tests to test_xet_download.py
